Refactored mocks into one file
