--- tests/mock_clock.py
@@ -1,3 +1,0 @@
-class MockClock:
-    def __init__(self):
-        self.now = 0
