+++ tests/mocks.py
@@ -1,4 +1,7 @@
-from .mock_clock import MockClock
+class MockClock:
+    def __init__(self):
+        self.now = 0
+
 
 class MockManager:
     def __init__(self):
